Storage: Refactor for fixing recreate Storage instance (#8701)
ref pingcap/tiflash#8695
