+++ dbms/src/Storages/DeltaMerge/StoragePool/StoragePool.h
@@ -1,4 +1,4 @@
-// Copyright 2023 PingCAP, Inc.
+// Copyright 2024 PingCAP, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,7 +17,7 @@
 #include <Common/CurrentMetrics.h>
 #include <Common/Logger.h>
 #include <Storages/BackgroundProcessingPool.h>
-#include <Storages/DeltaMerge/StoragePool_fwd.h>
+#include <Storages/DeltaMerge/StoragePool/StoragePool_fwd.h>
 #include <Storages/KVStore/Types.h>
 #include <Storages/Page/FileUsage.h>
 #include <Storages/Page/PageStorage_fwd.h>
@@ -38,48 +38,10 @@ class StoragePathPool;
 class PathPool;
 class StableDiskDelegator;
 class AsynchronousMetrics;
+} // namespace DB
 
-namespace DM
+namespace DB::DM
 {
-static constexpr std::chrono::seconds DELTA_MERGE_GC_PERIOD(60);
-
-class GlobalStoragePool : private boost::noncopyable
-{
-public:
-    using Clock = std::chrono::system_clock;
-    using Timepoint = Clock::time_point;
-    using Seconds = std::chrono::seconds;
-
-    GlobalStoragePool(const PathPool & path_pool, Context & global_ctx, const Settings & settings);
-
-    ~GlobalStoragePool();
-
-    void restore();
-
-    void shutdown();
-
-    friend class StoragePool;
-    friend class ::DB::AsynchronousMetrics;
-
-    // GC immediately
-    // Only used on dbgFuncMisc
-    bool gc();
-
-    FileUsageStatistics getLogFileUsage() const;
-
-private:
-    bool gc(const Settings & settings, bool immediately = false, const Seconds & try_gc_period = DELTA_MERGE_GC_PERIOD);
-
-private:
-    PageStoragePtr log_storage;
-    PageStoragePtr data_storage;
-    PageStoragePtr meta_storage;
-
-    std::atomic<Timepoint> last_try_gc_time = Clock::now();
-
-    Context & global_context;
-    BackgroundProcessingPool::TaskHandle gc_handle;
-};
 
 class StoragePool : private boost::noncopyable
 {
@@ -146,10 +108,8 @@ public:
     PageReaderPtr newLogReader(ReadLimiterPtr read_limiter, PageStorageSnapshotPtr & snapshot);
 
     PageReaderPtr newDataReader(ReadLimiterPtr read_limiter, bool snapshot_read, const String & tracing_id);
-    PageReaderPtr newDataReader(ReadLimiterPtr read_limiter, PageStorageSnapshotPtr & snapshot);
 
     PageReaderPtr newMetaReader(ReadLimiterPtr read_limiter, bool snapshot_read, const String & tracing_id);
-    PageReaderPtr newMetaReader(ReadLimiterPtr read_limiter, PageStorageSnapshotPtr & snapshot);
 
     // Register the clean up DMFiles callbacks to PageStorage.
     // The callbacks will be unregister when `shutdown` is called.
@@ -244,5 +204,4 @@ struct StorageSnapshot : private boost::noncopyable
 };
 
 
-} // namespace DM
-} // namespace DB
+} // namespace DB::DM
