+++ dbms/src/Storages/DeltaMerge/StoragePool/StoragePool_fwd.h
@@ -1,4 +1,4 @@
-// Copyright 2023 PingCAP, Inc.
+// Copyright 2024 PingCAP, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -27,5 +27,7 @@ using GlobalStoragePoolPtr = std::shared_ptr<GlobalStoragePool>;
 
 struct StorageSnapshot;
 using StorageSnapshotPtr = std::shared_ptr<StorageSnapshot>;
+
+static constexpr std::chrono::seconds DELTA_MERGE_GC_PERIOD(60);
 
 } // namespace DB::DM
