+++ dbms/src/Storages/Page/V3/Universal/tests/gtest_universal_page_storage_storage_pool.cpp
@@ -14,7 +14,7 @@
 
 #include <Interpreters/Context.h>
 #include <Poco/Logger.h>
-#include <Storages/DeltaMerge/StoragePool.h>
+#include <Storages/DeltaMerge/StoragePool/StoragePool.h>
 #include <Storages/Page/PageStorage.h>
 #include <Storages/Page/V3/Universal/UniversalPageIdFormatImpl.h>
 #include <Storages/Page/V3/Universal/UniversalPageStorage.h>
